pyproject.toml
[tool.poetry]
name = "patient-intake-agent"
version = "0.1.0"
description = ""
authors = ["avardaan <oyevardaan@gmail.com>"]
readme = "README.md"
packages = [{include = "patient_intake_agent"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
vocode = {extras = ["io"], version = "^0.1.110"}
python-dotenv = "^1.0.0"
redis = "^4.6.0"
twilio = "^8.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
